refactor(vame): duplicate functions
refactor duplicate functions, umap visualization and other fixes
